% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{noba.bench.log2}
\alias{noba.bench.log2}
\title{The randomized (benchmark) probe-level dataset, before preprocessed}
\format{A data matrix with 53880 rows (probes) and 192 columns (samples); column names ending with "E" or "V" differentiates whether a sample is endometrial or ovarian.}
\usage{
noba.bench.log2
}
\description{
The randomized probe-level dataset, non-control-probe-removed, no background adjusted and after logged 2.
}
\references{
\url{http://clincancerres.aacrjournals.org/content/20/13/3371.long}
}
\keyword{bench.data}

