% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{noba.nuhdata.log2}
\alias{noba.nuhdata.log2}
\title{The non-uniformly-handled probe-level dataset, before preprocessing}
\format{A data matrix with 53880 rows (probes) and 192 columns (samples).}
\usage{
noba.nuhdata.log2
}
\description{
The non-uniformly-handled probe-level dataset,
with control probes removed.
The expressions are on a log2 scale without background adjustment.
The sample IDs (the column names) ending with "E" or "V" are used to indicate
whether a sample is endometrial or ovarian tumor sample. There are
96 endometrial and 96 ovarian tumor samples.
}
\references{
\url{http://clincancerres.aacrjournals.org/content/20/13/3371.long}
}
\keyword{nuhdata}

