% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{noba.test.log2}
\alias{noba.test.log2}
\title{The non-randomized (test) probe-level dataset}
\format{A data matrix with 53880 rows (probes) and 192 columns (samples), column names ending with E/V are endometrial/ovarian samples.}
\usage{
noba.test.log2
}
\description{
The non-randomized probe-level dataset, non-control-probe-removed, no background adjusted and after logged 2.
}
\references{
\url{http://clincancerres.aacrjournals.org/content/20/13/3371.long}
}
\keyword{data}

