% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{noba.uhdata.log2.p10.psl}
\alias{noba.uhdata.log2.p10.psl}
\title{The uniformly-handled probe-set-level dataset}
\format{A data matrix with 3523 rows (probes) and 192 columns (samples).}
\usage{
noba.uhdata.log2.p10.psl
}
\description{
The uniformly-handled probe-set-level dataset.
It is probe-set summarized from the uniformly-handled probe-level dataset with
10 probes for each unique probe (\code{noba.uhdata.log2.p10}), using median.
}
\keyword{uhdata}

