% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{noba.uhdata.log2.p10}
\alias{noba.uhdata.log2.p10}
\title{The uniformly-handled probe-level dataset, 10 probes for each unique probe}
\format{A data matrix with 35230 rows (probes) and 192 columns (samples).}
\usage{
noba.uhdata.log2.p10
}
\description{
The uniformly-handled probe-level dataset
with 10 probes for each unique probe.
It is truncated from the uniformly-handled probe-level dataset
 (\code{noba.uhdata.log2.p10}), by keeping only the first 10 probes
 for each unique probe.
}
\keyword{uhdata}

