% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{unipbset}
\alias{unipbset}
\title{The estimated sample effect data}
\format{A data matrix with 53880 rows (probes) and 192 columns (samples), column names ending with E/V are endometrial/ovarian samples.}
\usage{
unipbset
}
\description{
The estimated sample effect data, derived from the randomized probe-level dataset.
}
\references{
\url{http://clincancerres.aacrjournals.org/content/20/13/3371.long}
}
\keyword{data}

