% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{noba.bench.log2.p10}
\alias{noba.bench.log2.p10}
\title{The randomized (benchmark) probe-level dataset, 10 probes for each unique probe}
\format{A data matrix with 35230 rows (probes) and 192 columns (samples), column names ending with E/V are endometrial/ovarian samples.}
\usage{
noba.bench.log2.p10
}
\description{
The randomized probe-level dataset, non-control-probe-removed, 10 probes for each unique probe, no background adjusted and after logged 2.
}
\references{
\url{http://clincancerres.aacrjournals.org/content/20/13/3371.long}
}
\keyword{data}

