% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{noba.test.log2.p10.psl}
\alias{noba.test.log2.p10.psl}
\title{The non-randomized (test) probe-set-level dataset, summarized from the non-randomized probe-level dataset with 10 probes for each unique probe}
\format{A data matrix with 3523 rows (probes) and 192 columns (samples), column names ending with E/V are endometrial/ovarian samples.}
\usage{
noba.test.log2.p10.psl
}
\description{
The non-randomized probe-set level dataset, non-control-probe-removed, no background adjusted and after logged 2, summarized from noba.test.log2.p10.
}
\references{
\url{http://clincancerres.aacrjournals.org/content/20/13/3371.long}
}
\keyword{data}

