% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{unipbset}
\alias{unipbset}
\title{Unique probe names}
\format{A list of two vectors one with length of 3523 and one with length of 6}
\usage{
unipbset
}
\description{
Unique probe names for non-control probes and negatively biological control probes from Agilent microarrays.
}
\keyword{probe.names}

